+++ internal/client/client.go
@@ -20,6 +20,8 @@ import (
 )
 
 type Client struct {
+	UsedLayer Layer // i.e. which auth layer are we using. Flags / Env Vars / Config File
+
 	CloudClient        *cloudv6.APIClient
 	AuthClient         *sdkgoauth.APIClient
 	CertManagerClient  *certmanager.APIClient
@@ -33,11 +35,7 @@ func appendUserAgent(userAgent string) string {
 	return fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), userAgent)
 }
 
-func newClient(name, pwd, token, hostUrl string) (*Client, error) {
-	if token == "" && (name == "" || pwd == "") {
-		return nil, errors.New("both token and at least one of username and password are empty")
-	}
-
+func newClient(name, pwd, token, hostUrl string) *Client {
 	clientConfig := cloudv6.NewConfiguration(name, pwd, token, hostUrl)
 	clientConfig.UserAgent = fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), clientConfig.UserAgent)
 	// Set Depth Query Parameter globally
@@ -62,21 +60,52 @@ func newClient(name, pwd, token, hostUrl string) (*Client, error) {
 	registryConfig.UserAgent = appendUserAgent(registryConfig.UserAgent)
 
 	return &Client{
-			CloudClient:        cloudv6.NewAPIClient(clientConfig),
-			AuthClient:         sdkgoauth.NewAPIClient(authConfig),
-			CertManagerClient:  certmanager.NewAPIClient(certManagerConfig),
-			PostgresClient:     postgres.NewAPIClient(postgresConfig),
-			MongoClient:        mongo.NewAPIClient(mongoConfig),
-			DataplatformClient: dataplatform.NewAPIClient(dpConfig),
-			RegistryClient:     registry.NewAPIClient(registryConfig),
-		},
-		nil
+		CloudClient:        cloudv6.NewAPIClient(clientConfig),
+		AuthClient:         sdkgoauth.NewAPIClient(authConfig),
+		CertManagerClient:  certmanager.NewAPIClient(certManagerConfig),
+		PostgresClient:     postgres.NewAPIClient(postgresConfig),
+		MongoClient:        mongo.NewAPIClient(mongoConfig),
+		DataplatformClient: dataplatform.NewAPIClient(dpConfig),
+		RegistryClient:     registry.NewAPIClient(registryConfig),
+	}
 }
 
 var once sync.Once
 var instance *Client
 
+func getFirstValidSource(layers []Layer) (values map[string]string, usedLayer Layer) {
+	for _, layer := range layers {
+		token := viper.GetString(layer.TokenKey)
+		username := viper.GetString(layer.UsernameKey)
+		password := viper.GetString(layer.PasswordKey)
+
+		if layer.TokenKey != "" && token != "" ||
+			layer.UsernameKey != "" && username != "" && layer.PasswordKey != "" && password != "" {
+			return map[string]string{
+				"token":    token,
+				"username": username,
+				"password": password,
+			}, layer
+		}
+	}
+	return nil, Layer{}
+}
+
+type Layer struct {
+	TokenKey    string
+	UsernameKey string
+	PasswordKey string
+	Help        string
+}
+
+var ConfigurationPriorityRules = []Layer{
+	{constants.ArgToken, "", "", fmt.Sprintf("Global Flags (--%s)", constants.ArgToken)},
+	{constants.EnvToken, constants.EnvUsername, constants.EnvPassword, fmt.Sprintf("Environment Variables (%s, %s, %s)", constants.EnvToken, constants.EnvUsername, constants.EnvPassword)},
+	{constants.CfgToken, constants.CfgUsername, constants.CfgPassword, fmt.Sprintf("Config file settings (%s, %s, %s)", constants.CfgToken, constants.CfgUsername, constants.CfgPassword)}, // Note: Username & Password are no longer generated in cfg file by `ionosctl login`, however we will keep this for backward compatibility.
+}
+
 // Get a client and possibly fail. Uses viper to get the credentials and API URL.
+// The returned client is guaranteed to have working credentials
 // Order:
 // Explicit flags ( e.g. --token )
 // Environment Variables ( e.g. IONOS_TOKEN )
@@ -99,41 +128,22 @@ func Get() (*Client, error) {
 
 		viper.AutomaticEnv()
 
-		// Credentials and API URL priority: command line arguments -> environment variables -> config file
-		token := viper.GetString(constants.ArgToken)
-		if token == "" {
-			token = viper.GetString(constants.EnvToken)
-		}
-		if token == "" {
-			token = viper.GetString(constants.CfgToken)
-		}
-
-		hostUrl := viper.GetString(constants.ArgServerUrl)
-		if hostUrl == "" {
-			hostUrl = viper.GetString(constants.EnvServerUrl)
-		}
-		if hostUrl == "" {
-			hostUrl = viper.GetString(constants.CfgServerUrl)
-		}
-
-		username := viper.GetString(constants.EnvUsername)
-		if username == "" {
-			username = viper.GetString(constants.CfgUsername)
-		}
-
-		password := viper.GetString(constants.EnvPassword)
-		if password == "" {
-			password = viper.GetString(constants.CfgPassword)
-		}
-
-		// Check if at least one authentication method is available
-		if token == "" && (username == "" || password == "") {
+		values, usedLayer := getFirstValidSource(ConfigurationPriorityRules)
+		if values == nil {
 			getClientErr = errors.Join(getClientErr, fmt.Errorf("not logged in: use either environment variables %s or %s and %s, or use `ionosctl login`", constants.EnvToken, constants.EnvUsername, constants.EnvPassword))
 			return
 		}
 
-		instance, err = newClient(username, password, token, hostUrl)
-		if err != nil {
+		instance = newClient(values["username"], values["password"], values["token"], values["serverUrl"])
+		instance.UsedLayer = usedLayer
+
+		// Check if at least one authentication method is available
+		if values["token"] == "" && (values["username"] == "" || values["password"] == "") {
+			getClientErr = errors.Join(getClientErr, fmt.Errorf("not logged in: use either environment variables %s or %s and %s, or use `ionosctl login`", constants.EnvToken, constants.EnvUsername, constants.EnvPassword))
+			return
+		}
+
+		if err := instance.TestCreds(); err != nil {
 			getClientErr = errors.Join(getClientErr, fmt.Errorf("failed creating client: %w", err))
 		}
 	})
@@ -142,8 +152,8 @@ func Get() (*Client, error) {
 }
 
 // Must gets the client obj or fatally dies
-// You can provide some optional custom error handlers as params
-// If error handlers not provided, dies with err message and exits with code 1
+// You can provide some optional custom error handlers as params. The err is sent to each error handler in order.
+// The default error handler is die.Die which exits with code 1 and violently terminates the program
 func Must(ehs ...func(error)) *Client {
 	client, err := Get()
 	if err != nil {
@@ -159,24 +169,29 @@ func Must(ehs ...func(error)) *Client {
 }
 
 // NewClient bypasses the singleton check, not recommended for normal use.
-func NewClient(name, pwd, token, hostUrl string) (*Client, error) {
+func NewClient(name, pwd, token, hostUrl string) *Client {
 	return newClient(name, pwd, token, hostUrl)
 }
 
 func TestCreds(user, pass, token string) error {
-	cl, err := newClient(user, pass, token, config.GetServerUrl())
-	if err != nil {
-		return fmt.Errorf("failed initializing client with credentials: %w", err)
-	}
+	cl := newClient(user, pass, token, constants.DefaultApiURL)
+	return cl.TestCreds()
+}
 
-	_, _, err = cl.CloudClient.DataCentersApi.DatacentersGet(context.Background()).Execute()
+func (c *Client) TestCreds() error {
+	_, _, err := c.CloudClient.DataCentersApi.DatacentersGet(context.Background()).Execute()
 	if err != nil {
 		usedScheme := "used token"
-		if token == "" {
-			usedScheme = fmt.Sprintf("used username '%s' and password", user)
+		if c.CloudClient.GetConfig().Token == "" {
+			usedScheme = fmt.Sprintf("used username '%s' and password", c.CloudClient.GetConfig().Username)
 		}
 		return fmt.Errorf("credentials test failed. %s: %w", usedScheme, err)
 	}
 
 	return nil
+}
+
+// IsTokenAuth returns true if a token is being used for authentication. Otherwise, username & password were used.
+func (c *Client) IsTokenAuth() bool {
+	return c.CloudClient.GetConfig().Token != ""
 }