+++ pkg/constants/constants.go
@@ -44,6 +44,12 @@ const (
 // Flag descriptions. Prefixed with "Desc" for easy find and replace
 const (
 	DescMaxResults = "The maximum number of elements to return"
+	// DescAuthenticationOrder explains auth order. Embed this in any auth-related commands
+	DescAuthenticationOrder = `ionosctl uses a layered approach for authentication, prioritizing sources in this order:
+  1. Global flags
+  2. Environment variables
+  3. Config file entries
+Within each layer, a token takes precedence over a username and password combination. For instance, if a token and a username/password pair are both defined in environment variables, ionosctl will prioritize the token. However, higher layers can override the use of a token from a lower layer. For example, username and password environment variables will supersede a token found in the config file.`
 )
 
 // legacy flags. TODO: Arg should be renamed to Flag.